+++ glamod_marine_processing/cli.py
@@ -169,6 +169,18 @@ class Options:
         self.run_jobs = click.option(
             "-run", "--run_jobs", is_flag=True, help="Run job scripts interactively."
         )
+        self.parallel_jobs = click.option(
+            "-parallel",
+            "--parallel_jobs",
+            is_flag=True,
+            help="Run job scripts interactively in parallel. This is mainly for BASTION machine.",
+        )
+        self.n_max_jobs = click.option(
+            "-n_max",
+            "--n_max_jobs",
+            default="12",
+            help="Maximum number of jobs running in parallel. Use only if parallel_jobs is True",
+        )
         self.level = click.option(
             "-l",
             "--level",